Backend sprint3

+++ src/main/java/com/liftoff/controllers/AdminUserController.java
@@ -1,8 +1,11 @@
 package com.liftoff.controllers;
 
+import com.liftoff.models.Role;
 import com.liftoff.models.User;
+import com.liftoff.models.data.RoleRepository;
 import com.liftoff.models.data.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,6 +13,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import javax.xml.bind.SchemaOutputResolver;
 import java.util.List;
 
 @Controller
@@ -18,6 +22,10 @@ public class AdminUserController {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private RoleRepository roleRepository;
+
 
     @GetMapping ("")
     public String displayManageUsersPage (Model model) {
@@ -35,25 +43,54 @@ public class AdminUserController {
 
     @GetMapping ("edit/{id}")
     public String displayEditUserPage(@PathVariable("id") Integer id, Model model){
-        List<User> listUsers = userRepository.findAll();
-        User user = userRepository.findById(id).get();
-        model.addAttribute("listUsers", listUsers);
-        model.addAttribute("title","Admin Portal: edit User");
-        model.addAttribute("button", "Save Changes");
-        model.addAttribute("user", user);
-        return "/account/manageUsers/edit";
+        if (!userRepository.existsById(id)) {
+            return "/500";
+        } else {
+                List<User> listUsers = userRepository.findAll();
+                List<Role> listRoles = roleRepository.findAll();
+                User user = userRepository.findById(id).get();
+                model.addAttribute("listUsers", listUsers);
+                model.addAttribute("listRoles", listRoles);
+                model.addAttribute("title", "Admin Portal: edit User");
+                model.addAttribute("button", "Save Changes");
+                model.addAttribute("user", user);
+                return "/account/manageUsers/edit";
+        }
+
+    }
+
+    @GetMapping("reset/{id}")
+    public String passwordReset (@PathVariable("id") Integer id) {
+        User user = userRepository.getById(id);
+        BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
+        String resetPassword = "welcome";
+        String encodedPassword = encoder.encode(resetPassword);
+        System.out.println("Resetting User: User found: " + user);
+
+        if (!userRepository.existsById(id)) {
+            return "/500";
+        }else{
+             user.setPassword(encodedPassword);
+        }
+            return "/account/manageUsers/index";
     }
 
     @GetMapping ("delete/{id}")
-    public String deleteUser (User user) {
-        userRepository.delete(user);
-        return "redirect:/account/manageUsers";
+    public String deleteUser (@PathVariable("id") Integer id) {
+        if (!userRepository.existsById(id)) {
+            return "/500";
+        }else{
+            User user = userRepository.getById(id);
+            userRepository.delete(user);
+        }
+            return "/account/manageUsers/index";
     }
 
     @PostMapping("save")
     public String saveUser (User user) {
         userRepository.save(user);
-        return "/account/manageUsers";
+        return "/account/manageUsers/index";
     }
+
 
 }
